Initial commit
The idea of this repository is to unify basic functionallity from all
supported languages. No source code will be source or type signatures
defined in this repo, it will only import the `*base` morloc modules from
other languages and re-export selection of them.

It is not required that all languages implement all exported functions.

The general signatures are defined in one place and is imported by each
`*base`, acting as implementations of a common interface.

                conventions
        ____________|_____________
        |          |       |      |
        cppbase rustbase pybase rbase
        |__________|_______|______|
                     |
                   base

The user can import select `*bases` or the entire base.

Now for naming, I've been calling things "base", which I don't really
like. I should convert over to the longer, but more lyrical, "prelude", or
perhaps "foundation". The idea is that this is one library that contains
all the basic stuff you want wrapped together and carefully tested for
compatibility.

Do I want an implicit prelude? No, certainly not. The power of morloc is
in free composition of things. Learning from the experience in Haskell,
there will be a strong desire in the future to implement alternative
preludes. The ideal function selection will depend on the domain.

A prelude may be focused on performance or on correctness (or both). It
may be maximalist or minimalist, inclusive (many languages) or exclusive
(few languages or just one). It may be carefully curated or free.